Clean up Logic code
