+++ src/diffusion_hash_inv/core/__init__.py
@@ -7,7 +7,7 @@ Common classes and utilities for diffusion_hash_inv.
 """
 
 from .base_calc import BaseCalc
-from .logger import BaseLogs, StepLogs, Metadata, Logs
+from .logger import BaseLogs, StepLogs, Metadata, Logs, MD5RoundTrace, MD5Step4Trace, LogDecorators
 from .rgb_type import RGB, RGBBinning
 from .byte2rgb import Byte2RGB
 from .fix_class_var import FreezeClassVar
@@ -18,6 +18,9 @@ __all__ = [
     "StepLogs",
     "Metadata",
     "Logs",
+    "MD5RoundTrace",
+    "MD5Step4Trace",
+    "LogDecorators",
     "RGB",
     "RGBBinning",
     "Byte2RGB",
